Separate servo lock functions
