+++ servo/lock.h
@@ -1,0 +1,10 @@
+//
+// Created by dfkdream on 22. 12. 9.
+//
+
+#ifndef SMART_DOORLOCK_LOCK_H
+#define SMART_DOORLOCK_LOCK_H
+
+void lock(int);
+
+#endif //SMART_DOORLOCK_LOCK_H
